Refactor authentication and notes services
