+++ src/app/guards/auth.guard.ts
@@ -1,7 +1,13 @@
 import { CanActivateFn } from '@angular/router';
 
 import { AuthService } from '../services/auth.service';
+import { inject } from '@angular/core';
 
 export const authGuard: CanActivateFn = (route, state) => {
-  return true;
+
+  if (inject(AuthService).isLoggedIn()) {
+    return true;
+  } else {
+    return false;
+  }
 };
