+++ src/app/services/auth.service.ts
@@ -1,4 +1,4 @@
-import { SocialAuthService, SocialUser } from '@abacritt/angularx-social-login';
+import { SocialAuthService, SocialUser, GoogleLoginProvider } from '@abacritt/angularx-social-login';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 
@@ -7,13 +7,23 @@ import { Injectable } from '@angular/core';
 })
 export class AuthService {
 
+  public accessToken: string = '';
+
   constructor(
     private authService: SocialAuthService,
     private httpClient: HttpClient
     ) { }
 
   public isLoggedIn(): boolean {
-    return false;
+    return !(!this.accessToken);
   }
-  
+
+  async getAccessToken(user: SocialUser): Promise<string> {
+    
+    await this.authService.getAccessToken(GoogleLoginProvider.PROVIDER_ID).then(accessToken => {
+      this.accessToken = accessToken;
+    });
+
+    return this.accessToken;
+  }
 }
