+++ src/app/services/notes.service.ts
@@ -1,12 +1,15 @@
 import { Injectable } from '@angular/core';
 import { Notes } from '../interfaces/notes';
 
+import { HttpClient } from '@angular/common/http';
+import { lastValueFrom } from 'rxjs';
+
 @Injectable({
   providedIn: 'root'
 })
 export class NotesService {
 
-  constructor() { }
+  constructor(private httpClient: HttpClient) { }
 
   getAllNotes(){
     return NOTES;
@@ -18,6 +21,26 @@ export class NotesService {
 
   createNote(note: Notes){
     NOTES.push(note);
+  }
+
+  async getPageNotes(accessToken: string, page: number = 0, notesPerPages: number = 3): Promise<string> {
+    const url = 'http://localhost:3300/auth';
+    const body = {
+      accessToken: accessToken,
+      page: page,
+      notesPerPages: notesPerPages,
+    };
+    const options = {
+      headers: {
+        'Content-Type': 'application/json'
+      },
+    };
+
+    try {
+      return await lastValueFrom(this.httpClient.post<string>(url, body, options));
+    } catch(error) {
+      return Promise.reject(error);
+    }
   }
 }
